+++ Android/NatShare/src/main/java/api/natsuite/natshare/SharePayload.java
@@ -11,13 +11,13 @@ import android.os.HandlerThread;
 import androidx.core.content.FileProvider;
 import android.util.Log;
 import android.webkit.MimeTypeMap;
-
 import com.unity3d.player.UnityPlayer;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
+import java.util.UUID;
 
 /**
  * NatShare
@@ -27,9 +27,12 @@ public final class SharePayload implements Payload {
 
     private final Intent intent;
     private final Callback completionHandler;
+    private final HandlerThread commitThread;
+    private final Handler handler;
     private final ArrayList<Uri> uris;
+    private final ArrayList<String> mimes;
+
     private static final String authority;
-    private String mime; // When sharing single item
 
     static { authority = UnityPlayer.currentActivity.getPackageName() + ".natshare"; }
 
@@ -38,6 +41,11 @@ public final class SharePayload implements Payload {
         this.intent = new Intent();
         this.completionHandler = completionHandler;
         this.uris = new ArrayList<>();
+        this.mimes = new ArrayList<>();
+        // Create handler
+        this.commitThread = new HandlerThread("SharePayload");
+        commitThread.start();
+        this.handler = new Handler(commitThread.getLooper());
         // Set intent params
         intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
@@ -46,52 +54,48 @@ public final class SharePayload implements Payload {
     @Override
     public void addText (final String text) {
         intent.putExtra(Intent.EXTRA_TEXT, text);
-        mime = "text/plain";
+        mimes.add("text/plain");
     }
 
     @Override
     public void addImage (final ByteBuffer jpegData) {
         // Read into managed memory
-        final byte[] buffer = new byte[jpegData.capacity()];
-        jpegData.clear();
+        final byte[] buffer = new byte[jpegData.limit()];
         jpegData.get(buffer);
         // Write to file
-        try {
-            File file = new File(UnityPlayer.currentActivity.getCacheDir(), "share." + System.nanoTime() + ".jpg");
-            FileOutputStream outputStream = new FileOutputStream(file);
-            outputStream.write(buffer);
-            outputStream.close();
-            addMedia(file.getAbsolutePath());
-            mime = "image/jpeg";
-        } catch (IOException ex) {
-            Log.e("NatSuite", "NatShare Error: SharePayload failed to commit image with error: " + ex);
-        }
+        handler.post(() -> {
+            try {
+                File file = new File(UnityPlayer.currentActivity.getCacheDir(), UUID.randomUUID().toString() + ".jpg");
+                FileOutputStream stream = new FileOutputStream(file);
+                stream.write(buffer);
+                stream.close();
+                addMedia(file.getAbsolutePath());
+            } catch (IOException ex) {
+                Log.e("NatSuite", "NatShare Error: SharePayload failed to commit image with error: " + ex);
+            }
+        });
     }
 
     @Override
-    public void addMedia (final String path) {
+    public synchronized void addMedia (final String path) {
         Uri uri = FileProvider.getUriForFile(UnityPlayer.currentActivity, authority, new File(path));
-        uris.add(uri);
         String extension = MimeTypeMap.getFileExtensionFromUrl(path);
-        mime = MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension);
+        String mime = MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension);
+        uris.add(uri);
+        mimes.add(mime);
     }
 
     @Override
     public void commit () {
-        final HandlerThread commitThread = new HandlerThread("SharePayload Commit Thread");
-        commitThread.start();
-        new Handler(commitThread.getLooper()).post(() -> {
+        // Check
+        handler.post(() -> {
             // Finalize intent
-            if (uris.size() > 1) {
-                intent.setAction(Intent.ACTION_SEND_MULTIPLE);
-                intent.putParcelableArrayListExtra(Intent.EXTRA_STREAM, uris);
-                intent.setType("*/*");
-            }
-            else if (uris.size() == 1) {
-                intent.setAction(Intent.ACTION_SEND);
+            intent.setAction(uris.size() > 1 ? Intent.ACTION_SEND_MULTIPLE : Intent.ACTION_SEND);
+            intent.setType(flattenMime(mimes));
+            if (uris.size() == 1)
                 intent.putExtra(Intent.EXTRA_STREAM, uris.get(0));
-                intent.setType(mime);
-            }
+            else if (uris.size() > 1)
+                intent.putParcelableArrayListExtra(Intent.EXTRA_STREAM, uris);
             // Start activity
             ShareReceiver.completionHandler = completionHandler;
             Intent receiver = new Intent(UnityPlayer.currentActivity, ShareReceiver.class);
@@ -111,5 +115,21 @@ public final class SharePayload implements Payload {
             final ComponentName clickedComponent = intent.getParcelableExtra(Intent.EXTRA_CHOSEN_COMPONENT);
             completionHandler.onCompletion(true);
         }
+    }
+
+    private static String flattenMime (ArrayList<String> mimes) {
+        // Check for single
+        if (mimes.size() == 0)
+            return "*/*";
+        else if (mimes.size() == 1)
+            return mimes.get(0);
+        // Find lowest common mime type
+        ArrayList<String> mediaTypes = new ArrayList<>();
+        for (String mime : mimes) {
+            String type = mime.split("/")[0];
+            if (!mediaTypes.contains(type))
+                mediaTypes.add(type);
+        }
+        return mediaTypes.size() > 1 ? "*/*" : mediaTypes.get(0) + "/*";
     }
 }