+++ Assets/NatSuite/Plugins/Managed/Sharing/PrintPayload.cs
@@ -26,7 +26,7 @@ namespace NatSuite.Sharing {
         /// Add text to the payload.
         /// </summary>
         public ISharePayload AddText (string text) {
-            payload?.AddText(text);
+            payload.AddText(text);
             return this;
         }
 
@@ -36,7 +36,7 @@ namespace NatSuite.Sharing {
         /// </summary>
         /// <param name="image">Image to be added to the gallery.</param>
         public ISharePayload AddImage (Texture2D image) {
-            payload?.AddImage(image);
+            payload.AddImage(image);
             return this;
         }
 
@@ -45,14 +45,14 @@ namespace NatSuite.Sharing {
         /// </summary>
         /// <param name="path">Path to local media file to be shared.</param>
         public ISharePayload AddMedia (string uri) {
-            payload?.AddMedia(uri);
+            payload.AddMedia(uri);
             return this;
         }
 
         /// <summary>
         /// Commit the payload and return whether payload was successfully shared.
         /// </summary>
-        public Task<bool> Commit () => payload?.Commit();
+        public Task<bool> Commit () => payload.Commit();
         #endregion
 
         private readonly ISharePayload payload;