+++ Assets/NatSuite/Plugins/Managed/Sharing/SharePayload.cs
@@ -24,7 +24,7 @@ namespace NatSuite.Sharing {
         /// Add plain text.
         /// </summary>
         public ISharePayload AddText (string text) {
-            payload?.AddText(text);
+            payload.AddText(text);
             return this;
         }
 
@@ -34,7 +34,7 @@ namespace NatSuite.Sharing {
         /// </summary>
         /// <param name="image">Image to be shared.</param>
         public ISharePayload AddImage (Texture2D image) {
-            payload?.AddImage(image);
+            payload.AddImage(image);
             return this;
         }
 
@@ -43,14 +43,14 @@ namespace NatSuite.Sharing {
         /// </summary>
         /// <param name="path">Path to local media file to be shared.</param>
         public ISharePayload AddMedia (string path) {
-            payload?.AddMedia(path);
+            payload.AddMedia(path);
             return this;
         }
 
         /// <summary>
         /// Commit the payload and return whether payload was successfully shared.
         /// </summary>
-        public Task<bool> Commit () => payload?.Commit();
+        public Task<bool> Commit () => payload.Commit();
         #endregion
 
         private readonly ISharePayload payload;