+++ Assets/NatSuite/Plugins/Managed/Sharing/Changelog.md
@@ -1,5 +1,6 @@
 ## NatShare 1.2.3
-+ Fix compiler errors when building with IL2CPP backend on Android.
++ Fixed compiler errors when building with IL2CPP backend on Android.
++ Fixed 'No apps can perform this action' error when sharing text on Android.
 + Reduced time taken when adding images to payloads.
 
 ## NatShare 1.2.2